+++ packages/wepy/src/app.js
@@ -1,4 +1,3 @@
-
 let RequestMQ = {
     map: {},
     mq: [],
@@ -29,7 +28,7 @@ let RequestMQ = {
                     me.next();
                 }
                 this.running.push(obj.t);
-                return wx.request_bak(obj);
+                return wx.request(obj);
         }
     },
     request (obj) {
@@ -48,14 +47,39 @@ let RequestMQ = {
 
 export default class {
 
-    init () {
-        this.addPromise();
-        this.hackRequest();
+    $addons = {};
+
+    $interceptors = {};
+
+
+
+    init (wepy) {
+        this.initAPI(wepy);
         this.$wxapp = getApp();
     }
 
-    addPromise () {
 
+    use (addon, ...args) {
+        if (typeof(addon) === 'string' && this[addon]) {
+            this.$addons[addon] = 1;
+            this[addon](args);
+        } else {
+            this.$addons[addon.name] = new addon(args);
+        }
+    }
+
+    intercept (api, provider) {
+        this.$interceptors[api] = provider;
+    }
+
+    promisify () {
+    }
+
+    requestfix () {
+    }
+
+    initAPI (wepy) {
+        var self = this;
         let noPromiseMethods = {
             stopRecord: true,
             pauseVoice: true,
@@ -71,55 +95,71 @@ export default class {
             stopPullDownRefresh: true
         };
         Object.keys(wx).forEach((key) => {
-            if (!noPromiseMethods[key] && key.substr(0, 2) !== 'on' && key !== 'request' && !(/\w+Sync$/.test(key))) {
-                wx[key + '_bak'] = wx[key];
-                Object.defineProperty(wx, key, {
+            if (!noPromiseMethods[key] && key.substr(0, 2) !== 'on' && !(/\w+Sync$/.test(key))) {
+                Object.defineProperty(wepy, key, {
                     get () {
                         return (obj) => {
                             obj = obj || {};
-                            //obj = (typeof(obj) === 'string') ? {url: obj} : obj;
-                            return new Promise((resolve, reject) => {
-                                obj.success = resolve;
-                                obj.fail = (res) => {
-                                    if (res && res.errMsg) {
-                                        reject(new Error(res.errMsg));
+                            if (self.$interceptors[key] && self.$interceptors[key].config) {
+                                let rst = self.$interceptors[key].config.call(self, obj);
+                                if (rst === false) {
+                                    if (self.$addons.promisify) {
+                                        return Promise.reject('aborted by interceptor');
                                     } else {
-                                        reject(res);
+                                        obj.fail && obj.fail('aborted by interceptor');
+                                        return;
                                     }
                                 }
-                                wx[key + '_bak'](obj);
-                            });
+                                obj = rst;
+                            }
+                            if (key === 'request') {
+                                obj = (typeof(obj) === 'string') ? {url: obj} : obj;
+                            }
+                            if (self.$addons.promisify) {
+                                return new Promise((resolve, reject) => {
+                                    let bak = {};
+                                    ['fail', 'success', 'complete'].forEach((k) => {
+                                        bak[k] = obj[k];
+                                        obj[k] = (res) => {
+                                            if (self.$interceptors[key] && self.$interceptors[key][k]) {
+                                                res = self.$interceptors[key][k].call(self, res);
+                                            }
+                                            if (k === 'success')
+                                                resolve(res)
+                                            else if (k === 'fail')
+                                                reject(res);
+                                        };
+                                    });
+                                    if (self.$addons.requestfix && key === 'request') {
+                                        RequestMQ.request(obj);
+                                    } else
+                                        wx[key](obj);
+                                });
+                            } else {
+                                let bak = {};
+                                ['fail', 'success', 'complete'].forEach((k) => {
+                                    bak[k] = obj[k];
+                                    obj[k] = (res) => {
+                                        if (self.$interceptors[key] && self.$interceptors[key][k]) {
+                                            res = self.$interceptors[key][k].call(self, res);
+                                        }
+                                        bak[k] && bak[k].call(self, res);
+                                    };
+                                });
+                                if (self.$addons.requestfix && key === 'request') {
+                                    RequestMQ.request(obj);
+                                } else
+                                    wx[key](obj);
+                            }
                         };
                     }
                 });
+            } else {
+                Object.defineProperty(wepy, key, {
+                    get () { return (obj) => wx[key](obj) }
+                });
             }
         });
+
     }
-
-
-    hackRequest () {
-        
-        wx['request_bak'] = wx['request'];
-        Object.defineProperty(wx, 'request', {
-            get () {
-                return (obj) => {
-                    obj = obj || {};
-                    obj = (typeof(obj) === 'string') ? {url: obj} : obj;
-                    return new Promise((resolve, reject) => {
-                        obj.success = resolve;
-                        obj.fail = (res) => {
-                            if (res && res.errMsg) {
-                                reject(new Error(res.errMsg));
-                            } else {
-                                reject(res);
-                            }
-                        }
-                        RequestMQ.request(obj);
-                    })
-                };
-            }
-        });
-    }
-
-
 }